Add zoom folder check and add directory Fix empty command zoom bug by returning exception
