+++ src/main/java/seedu/duke/storage/Storage.java
@@ -3,6 +3,7 @@ package seedu.duke.storage;
 import seedu.duke.food.WhatIAteList;
 import seedu.duke.journal.CollectionOfEntries;
 import seedu.duke.journal.CollectionOfNotes;
+import seedu.duke.logger.ClickLogger;
 import seedu.duke.task.TaskList;
 
 import java.io.BufferedReader;
@@ -15,6 +16,7 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.logging.Level;
 
 public class Storage {
 
@@ -44,13 +46,25 @@ public class Storage {
     public Storage() throws IOException {
     }
 
-    //@author nigel
-    public static void checkAndAddDirectory(String folderName) throws IOException {
+    /**
+     * Creates directory if directory folderName is not found.
+     * access modifier left empty for working in Storage.
+     * @param folderName folder name to check
+     * @throws IOException case where directory not found
+     *
+     * @author ngnigel99
+     */
+    static void checkAndAddDirectory(String folderName) throws IOException {
+        assert folderName.contains("/") : "Please follow format [DIR_NAME]/";   //checks syntax of constant
         String home = new File("").getAbsolutePath() + '/';
         File dirCheck = new File(home + folderName);
+        ClickLogger.getNewLogger().log(Level.INFO,  "checking if directory exists!");
         if (dirCheck.isDirectory()) {
+            ClickLogger.getNewLogger().log(Level.INFO, "nice, directory found!");
             return;
         }
+        ClickLogger.getNewLogger().log(Level.CONFIG, "not nice, no directory found,"
+                +     "creating  new directory");
         System.out.println("Hey, I didn't find directory " + folderName);
         System.out.println("adding " + folderName + " into repository...");
         Files.createDirectories(Paths.get(home + folderName));
